requirements.txt
azure-storage-blob==12.17.0
python-magic==0.4.27
pypdf==3.15.4
requests==2.31.0
selenium==4.11.2
pandas==2.0.3
beautifulsoup4==4.12.2
openai==0.27.9
tiktoken==0.4.0
PyMuPDF==1.23.2
pyodbc==4.0.39
Flask-API==3.1
Flask-RESTful==0.3.10
python-magic==0.4.27